# swagger.json
{
 "swagger": "2.0",
 "info": {
   "contact": {},
   "title": "Full-Auth",
   "version": "1.0.0"
 },
 "basePath": "/api/v1",
 "paths": {
   "/auth/reset-password": {
     "post": {
       "consumes": [
         "application/json"
       ],
       "parameters": [
         {
           "in": "query",
           "name": "token",
           "type": "string"
         },
         {
           "in": "query",
           "name": "email",
           "type": "string"
         },
         {
           "in": "body",
           "name": "body",
           "schema": {
             "properties": {
               "email": {
                 "example": "fethi-dev@gmail.com",
                 "type": "string"
               },
               "password": {
                 "example": "password",
                 "type": "string"
               },
               "token": {
                 "example": "06027be8738ea7d287c112e1652d1bc940e34de0b3a3da9e23a5a6598f2518e7c38f96767ab4859cee96ba70ce908e94d7b0a8ae729a455a60b4a428464825d5a09852885ecc",
                 "type": "string"
               }
             },
             "type": "object"
           }
         }
       ],
       "responses": {
         "200": {
           "description": ""
         }
       },
       "description": "ResetPassword",
       "operationId": "resetpassword",
       "summary": "ResetPassword"
     }
   },
...
   "/auth/update": {
     "patch": {
       "consumes": [
         "application/json"
       ],
       "parameters": [
         {
           "in": "body",
           "name": "body",
           "schema": {
             "properties": {},
             "type": "object"
           }
         }
       ],
       "responses": {
         "200": {
           "description": ""
         }
       },
       "description": "UpdateUser",
       "operationId": "updateuser",
       "summary": "UpdateUser"
     }
   },
   "/auth/verify-email": {
     "get": {
       "consumes": [
         "application/json"
       ],
       "parameters": [
         {
           "in": "body",
           "name": "body",
           "schema": {
             "properties": {
               "email": {
                 "example": "fethi-dev@gmail.com",
                 "type": "string"
               },
               "otp": {
                 "example": "264317",
                 "type": "string"
               }
             },
             "type": "object"
           }
         }
       ],
       "responses": {
         "200": {
           "description": ""
         }
       },
       "description": "VerifyEmail",
       "operationId": "verifyemail",
       "summary": "VerifyEmail"
     }
   },
...
 },
...
}